repository interfaces moved to data package
