+++ app/src/main/java/com/spaceapp/core/di/RepositoryModule.kt
@@ -17,6 +17,20 @@ import com.spaceapp.data.datasource.remote.space_news.SpaceNewsRemoteDataSource
 import com.spaceapp.data.datasource.remote.weather_condition.WeatherConditionRemoteDataSource
 import com.spaceapp.data.datasource.remote.where_is_the_iss.WhereIsTheIssRemoteDataSource
 import com.spaceapp.data.repository.*
+import com.spaceapp.data.repository.auth.AuthRepository
+import com.spaceapp.data.repository.auth.AuthRepositoryImpl
+import com.spaceapp.data.repository.iss.WhereIsTheIssRepository
+import com.spaceapp.data.repository.iss.WhereIsTheIssRepositoryImpl
+import com.spaceapp.data.repository.nasa.NasaRepository
+import com.spaceapp.data.repository.nasa.NasaRepositoryImpl
+import com.spaceapp.data.repository.people_in_space.PeopleInSpaceRepository
+import com.spaceapp.data.repository.people_in_space.PeopleInSpaceRepositoryImpl
+import com.spaceapp.data.repository.space_news.SpaceNewsRepository
+import com.spaceapp.data.repository.space_news.SpaceNewsRepositoryImpl
+import com.spaceapp.data.repository.space_objects.SpaceObjectsRepository
+import com.spaceapp.data.repository.space_objects.SpaceObjectsRepositoryImpl
+import com.spaceapp.data.repository.weather.WeatherConditionRepository
+import com.spaceapp.data.repository.weather.WeatherConditionRepositoryImpl
 import com.spaceapp.domain.repository.*
 import dagger.Module
 import dagger.Provides
