+++ app/src/main/java/com/spaceapp/data/repository/auth/AuthRepository.kt
@@ -1,4 +1,4 @@
-package com.spaceapp.domain.repository
+package com.spaceapp.data.repository.auth
 
 import com.google.android.gms.tasks.Task
 import com.google.firebase.auth.AuthResult
