+++ app/src/main/java/com/spaceapp/data/repository/iss/WhereIsTheIssRepositoryImpl.kt
@@ -1,11 +1,10 @@
-package com.spaceapp.data.repository
+package com.spaceapp.data.repository.iss
 
 import com.spaceapp.data.datasource.local.where_is_the_iss.WhereIsTheIssLocalDataSource
 import com.spaceapp.data.datasource.remote.where_is_the_iss.WhereIsTheIssRemoteDataSource
 import com.spaceapp.data.mappers.toIss
 import com.spaceapp.data.mappers.toIssEntity
 import com.spaceapp.domain.model.where_is_the_iss.Iss
-import com.spaceapp.domain.repository.WhereIsTheIssRepository
 import javax.inject.Inject
 
 class WhereIsTheIssRepositoryImpl @Inject constructor(
