+++ app/src/main/java/com/spaceapp/data/repository/nasa/NasaRepository.kt
@@ -1,4 +1,4 @@
-package com.spaceapp.domain.repository
+package com.spaceapp.data.repository.nasa
 
 import com.spaceapp.domain.model.apod.Apod
 import com.spaceapp.domain.model.mars_photos.MarsPhoto
