+++ app/src/main/java/com/spaceapp/data/repository/people_in_space/PeopleInSpaceRepositoryImpl.kt
@@ -1,11 +1,10 @@
-package com.spaceapp.data.repository
+package com.spaceapp.data.repository.people_in_space
 
 import com.spaceapp.data.datasource.local.people_in_space.PeopleInSpaceLocalDataSource
 import com.spaceapp.data.datasource.remote.people_in_space.PeopleInSpaceRemoteDataSource
 import com.spaceapp.data.mappers.toPeopleInSpace
 import com.spaceapp.data.mappers.toPeopleInSpaceEntity
 import com.spaceapp.domain.model.people_in_space.PeopleInSpace
-import com.spaceapp.domain.repository.PeopleInSpaceRepository
 import javax.inject.Inject
 
 class PeopleInSpaceRepositoryImpl @Inject constructor(
