+++ app/src/main/java/com/spaceapp/data/repository/space_objects/SpaceObjectsRepository.kt
@@ -1,4 +1,4 @@
-package com.spaceapp.domain.repository
+package com.spaceapp.data.repository.space_objects
 
 import android.content.Context
 import com.spaceapp.domain.model.explore_galaxy_data.SpaceObject
