+++ app/src/main/java/com/spaceapp/data/repository/space_objects/SpaceObjectsRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.spaceapp.data.repository
+package com.spaceapp.data.repository.space_objects
 
 import android.content.Context
 import com.spaceapp.data.datasource.local.explore_galaxy_data.ExploreGalaxyDataSource
@@ -7,7 +7,6 @@ import com.spaceapp.data.mappers.toGlossary
 import com.spaceapp.data.mappers.toSpaceObject
 import com.spaceapp.domain.model.explore_galaxy_data.SpaceObject
 import com.spaceapp.domain.model.glossary_data.Glossary
-import com.spaceapp.domain.repository.SpaceObjectsRepository
 import javax.inject.Inject
 
 class SpaceObjectsRepositoryImpl @Inject constructor(
