+++ app/src/main/java/com/spaceapp/data/repository/weather/WeatherConditionRepositoryImpl.kt
@@ -1,11 +1,10 @@
-package com.spaceapp.data.repository
+package com.spaceapp.data.repository.weather
 
 import com.spaceapp.data.datasource.local.weather_condition.WeatherConditionLocalDataSource
 import com.spaceapp.data.datasource.remote.weather_condition.WeatherConditionRemoteDataSource
 import com.spaceapp.data.mappers.toWeatherCondition
 import com.spaceapp.data.mappers.toWeatherConditionEntity
 import com.spaceapp.domain.model.weather_condition.WeatherCondition
-import com.spaceapp.domain.repository.WeatherConditionRepository
 import javax.inject.Inject
 
 class WeatherConditionRepositoryImpl @Inject constructor(
