+++ app/src/main/java/com/spaceapp/domain/usecase/apod/GetApodFromNetworkUseCase.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 import com.spaceapp.core.common.Result
-import com.spaceapp.domain.repository.NasaRepository
+import com.spaceapp.data.repository.nasa.NasaRepository
 import okio.IOException
 
 class GetApodFromNetworkUseCase @Inject constructor(private val nasaRepository: NasaRepository) {
