+++ app/src/main/java/com/spaceapp/domain/usecase/auth/ForgotPasswordUseCase.kt
@@ -4,7 +4,7 @@ import com.google.android.gms.tasks.Task as gmsTask
 import com.huawei.hmf.tasks.Task as hmsTask
 import com.spaceapp.core.common.TaskResult
 import com.spaceapp.domain.model.auth.ForgotPassword
-import com.spaceapp.domain.repository.AuthRepository
+import com.spaceapp.data.repository.auth.AuthRepository
 import com.spaceapp.domain.utils.ERROR
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
