+++ app/src/main/java/com/spaceapp/domain/usecase/mars_photo/AddMarsPhotoToDatabaseUseCase.kt
@@ -1,7 +1,7 @@
 package com.spaceapp.domain.usecase.mars_photo
 
 import com.spaceapp.domain.model.mars_photos.MarsPhoto
-import com.spaceapp.domain.repository.NasaRepository
+import com.spaceapp.data.repository.nasa.NasaRepository
 import javax.inject.Inject
 
 class AddMarsPhotoToDatabaseUseCase @Inject constructor(private val nasaRepository: NasaRepository) {
