+++ app/src/main/java/com/spaceapp/domain/usecase/mars_photo/ClearMarsPhotoDatabaseUseCase.kt
@@ -1,6 +1,6 @@
 package com.spaceapp.domain.usecase.mars_photo
 
-import com.spaceapp.domain.repository.NasaRepository
+import com.spaceapp.data.repository.nasa.NasaRepository
 import javax.inject.Inject
 
 class ClearMarsPhotoDatabaseUseCase @Inject constructor(private val nasaRepository: NasaRepository) {
