+++ app/src/main/java/com/spaceapp/domain/usecase/people_in_space/ClearPeopleInSpaceDatabaseUseCase.kt
@@ -1,6 +1,6 @@
 package com.spaceapp.domain.usecase.people_in_space
 
-import com.spaceapp.domain.repository.PeopleInSpaceRepository
+import com.spaceapp.data.repository.people_in_space.PeopleInSpaceRepository
 import javax.inject.Inject
 
 class ClearPeopleInSpaceDatabaseUseCase @Inject constructor(private val peopleInSpaceRepository: PeopleInSpaceRepository) {
