+++ app/src/main/java/com/spaceapp/domain/usecase/space_news/AddSpaceNewsToDatabaseUseCase.kt
@@ -2,7 +2,7 @@ package com.spaceapp.domain.usecase.space_news
 
 import com.spaceapp.domain.model.space_news.NewsImage
 import com.spaceapp.domain.model.space_news.SpaceNews
-import com.spaceapp.domain.repository.SpaceNewsRepository
+import com.spaceapp.data.repository.space_news.SpaceNewsRepository
 import javax.inject.Inject
 
 class AddSpaceNewsToDatabaseUseCase @Inject constructor(private val spaceNewsRepository: SpaceNewsRepository) {
