+++ app/src/main/java/com/spaceapp/domain/usecase/space_news/ClearSpaceNewsDatabaseUseCase.kt
@@ -1,6 +1,6 @@
 package com.spaceapp.domain.usecase.space_news
 
-import com.spaceapp.domain.repository.SpaceNewsRepository
+import com.spaceapp.data.repository.space_news.SpaceNewsRepository
 import javax.inject.Inject
 
 class ClearSpaceNewsDatabaseUseCase @Inject constructor(private val spaceNewsRepository: SpaceNewsRepository) {
