+++ app/src/main/java/com/spaceapp/domain/usecase/space_news/GetSpaceNewsFromDatabaseUseCase.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 import com.spaceapp.core.common.Result
-import com.spaceapp.domain.repository.SpaceNewsRepository
+import com.spaceapp.data.repository.space_news.SpaceNewsRepository
 import com.spaceapp.domain.utils.ERROR
 
 class GetSpaceNewsFromDatabaseUseCase @Inject constructor(private val spaceNewsRepository: SpaceNewsRepository) {
