+++ app/src/main/java/com/spaceapp/domain/usecase/weather_condition/AddWeatherToDatabaseUseCase.kt
@@ -1,7 +1,7 @@
 package com.spaceapp.domain.usecase.weather_condition
 
 import com.spaceapp.domain.model.weather_condition.WeatherCondition
-import com.spaceapp.domain.repository.WeatherConditionRepository
+import com.spaceapp.data.repository.weather.WeatherConditionRepository
 import javax.inject.Inject
 
 class AddWeatherToDatabaseUseCase @Inject constructor(private val weatherRepository: WeatherConditionRepository) {
