+++ app/src/main/java/com/spaceapp/domain/usecase/where_is_the_iss/UpdateIssPositionUseCase.kt
@@ -1,7 +1,7 @@
 package com.spaceapp.domain.usecase.where_is_the_iss
 
 import com.spaceapp.domain.model.where_is_the_iss.Iss
-import com.spaceapp.domain.repository.WhereIsTheIssRepository
+import com.spaceapp.data.repository.iss.WhereIsTheIssRepository
 import javax.inject.Inject
 
 class UpdateIssPositionUseCase @Inject constructor(private val issRepository: WhereIsTheIssRepository) {
